Directory Management
Moved new login.tsx into correct directory and replaced the default version.
