--- login.tsx
@@ -1,74 +1,0 @@
-import React, { useState } from 'react';
-import { Text, View, TextInput, StyleSheet, Button, Alert } from 'react-native';
-
-export default function LoginScreen() {
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-  const [savedUsername, setSavedUsername] = useState('');
-  const [savedPassword, setSavedPassword] = useState('');
-
-  const saveCredentials = () => {
-    setSavedUsername(username);
-    setSavedPassword(password);
-    Alert.alert('Saved', 'Your username and password have been saved.');
-  };
-
-  const login = () => {
-    if (username === savedUsername && password === savedPassword) {
-      Alert.alert('Welcome', 'You are now logged in!');
-    } else {
-      Alert.alert('Error', 'Incorrect username or password.');
-    }
-  };
-
-  return (
-    <View style={styles.container}>
-      <Text style={styles.label}>Username:</Text>
-      <TextInput
-        style={styles.input}
-        placeholder="Enter your username"
-        onChangeText={setUsername}
-        value={username}
-      />
-      <Text style={styles.label}>Password:</Text>
-      <TextInput
-        style={styles.input}
-        placeholder="Enter your password"
-        secureTextEntry={true}
-        onChangeText={setPassword}
-        value={password}
-      />
-      <View style={styles.buttonContainer}>
-        <Button title="Save" onPress={saveCredentials} />
-        <Button title="Login" onPress={login} />
-      </View>
-    </View>
-  );
-}
-
-const styles = StyleSheet.create({
-  container: {
-    flex: 1,
-    padding: 20,
-    justifyContent: 'center',
-    backgroundColor: '#f5f5f5',
-  },
-  label: {
-    fontSize: 16,
-    marginBottom: 5,
-    color: '#333',
-  },
-  input: {
-    borderWidth: 1,
-    borderColor: '#ccc',
-    padding: 10,
-    marginBottom: 15,
-    borderRadius: 5,
-    backgroundColor: '#fff',
-  },
-  buttonContainer: {
-    flexDirection: 'row',
-    justifyContent: 'space-between',
-    marginTop: 10,
-  },
-});
